+++ ListEngine/ListEngine/Detail Presentation/DetailViewModel.swift
@@ -1,0 +1,16 @@
+//
+//  DetailViewModel.swift
+//  ListEngine
+//
+//  Created by Afsal on 03/05/2024.
+//
+
+import Foundation
+
+public struct DetailViewModel {
+  public let name: String
+  public let country: String
+  public let code: String
+  public let state: String?
+  public let webpage: String?
+}
