+++ ListEngine/ListEngine/Feed API/RemoteFeedLoader.swift
@@ -11,17 +11,19 @@ public final class RemoteFeedLoader: FeedLoader {
   private let url: URL
   private let client: HTTPClient
   
+  public init(url: URL, client: HTTPClient) {
+    self.url = url
+    self.client = client
+  }
+}
+ 
+extension RemoteFeedLoader {
   public enum Error: Swift.Error {
     case connectivity
     case invalidData
   }
   
   public typealias Result = FeedLoader.Result
-  
-  public init(url: URL, client: HTTPClient) {
-    self.url = url
-    self.client = client
-  }
   
   public func load(completion: @escaping (Result) -> Void) {
     client.get(from: url) { [weak self] result in
